Separate extension methods. Move Windows.Media related conversion into SharedModel project.
