+++ Source/HelixToolkit.SharpDX.Shared/Extensions/BoundingSphereExtensions.cs
@@ -50,9 +50,8 @@ namespace HelixToolkit.Wpf.SharpDX
             {
                 //We are doing a relative distance comparison to find the maximum distance
                 //from the center of our sphere.
-                float distance;
                 var p = points[i];
-                Vector3.DistanceSquared(ref center, ref p, out distance);
+                Vector3.DistanceSquared(ref center, ref p, out float distance);
 
                 if (distance > radius)
                     radius = distance;
