+++ Source/HelixToolkit.SharpDX.SharedModel/Extensions/CommonExtensions.cs
