+++ Source/HelixToolkit.SharpDX.SharedModel/HelixToolkit.SharpDX.SharedModel.projitems
@@ -46,7 +46,9 @@
     <Compile Include="$(MSBuildThisFileDirectory)Element3D\PostEffects\PostEffectMeshXRayGrid.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Element3D\ScreenSpacedGroup3D.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Element3D\TransformManipulator3D.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Extensions\CommonExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\MaterialExtension.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Extensions\Media3DExtension.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Light3D\ShadowMap3D.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Material\ColorStripeMaterial.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Material\DiffuseMaterial.cs" />
